pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-scraper-ai"
version = "0.1.0"
description = "CLI-first job scraping scaffold for Indeed and future sources"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32,<3",
    "lxml>=5.2,<6",
    "pydantic>=2.7,<3",
    "python-dotenv>=1.0,<2",
    "tenacity>=8.3,<9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-cov>=5,<6",
]

[project.scripts]
job-scraper-ai = "job_scraper_ai.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
